# Improve inflate performance for repetitive data