I noticed that the new inflate implementation was slower when it comes to repeating data, like for example a bunch of zeroes.
This is relevant for things like images and voxel data, where we often see long runs of repeating data.

To improve this I used memcopies of iteratively doubling size for repeating data like this.
This is a fairly common approach and was also present in the old inflate implementation.

### Benchmark
<details>

<summary>Code</summary>

```zig
const std = @import("std");

pub fn deflate(allocator: std.mem.Allocator, data: []const u8) ![]u8 {
	var result = std.ArrayList(u8).init(allocator);
	var comp = try std.compress.flate.compressor(result.writer(), .{});
	_ = try comp.write(data);
	try comp.finish();
	return try result.toOwnedSlice();
}

pub fn inflateTo(buf: []u8, data: []const u8) !usize {
	var streamIn = std.io.fixedBufferStream(data);
	var decomp = std.compress.flate.decompressor(streamIn.reader());
	var streamOut = std.io.fixedBufferStream(buf);
	try decomp.decompress(streamOut.writer());
	return streamOut.getWritten().len;
}

const file: []const u8 = @embedFile("data.bin");
var in: [file.len]u8 = undefined;
var out: [file.len]u8 = undefined;

pub fn main() !void {
	@memcpy(&in, file);
	const t1 = std.time.nanoTimestamp();
	const compr = try deflate(std.heap.page_allocator, &in);
	defer std.heap.page_allocator.free(compr);
	const t2 = std.time.nanoTimestamp();
	_ = try inflateTo(&out, compr);
	const t3 = std.time.nanoTimestamp();
	std.log.err("{} {} {}", .{t2 - t1, t3 - t2, compr.len});
	for(&in, &out) |a, b| {
		std.debug.assert(a == b);
	}
}
```

</details>

data | master | this PR | speedup
-- | -- | -- | --
1 GB filled with zeroes | 1.5 s (1.0 s before the deflate rewrite) | 0.17 s | 8.8×
some uncompressed images with a small palette | 1.05 ms | 0.46 ms | 2.3×
zig/src folder put into an uncompressed tar archive | 273 ms | 269 ms | 1.0×