@ianic 2024-02-28
Thanks! Looks great.

I was never testing with that kind of data. 
I'll do some more benchmarks because I still find previous v1 implementation 1.6 times faster.
I got this timings by running above code:
```
v1:     132141572
this:   215850413
std:   1519825334
```