@IntegratedQuantum 2024-02-28
> I still find previous v1 implementation 1.6 times faster.

Right, I can also measure this now. I must have hallucinated an extra digit in my tests yesterday, it's 0.1 s for me, not 1.0 s.

I looked at the previous implementation, and honestly I have no clue why it is faster.