@ianic 2024-02-28
No this is valuable improvement. When I'm referring to v1 I think about implementation we got few weeks ago. This commit improves current stdlib implementation. Although we are still behind v1 for this particular case (huge empty file). In many other cases we are now [better](https://github.com/ianic/flate) than v1. In average 1.3 times for decompression. 

I did few benchmarks for this empty files and have some findings but still need some time to summarize them.