@ianic 2024-02-28
For this data (1G empty) if I replace `@memcpy` with `std.mem.copyForward` (as used in v1) I got better result 144ms instead of 215ms. 

Empty file is compressed as series of back references of length 258 (max length) and distance 1. That tells decompressor to go 1 byte back and copy 258 bytes from that position to the current position. In 1G file there are 4_161_790 such back references.
That backrefernce runs [loop](https://github.com/ziglang/zig/blob/9410b11ca663231e367e3adfd668979c4b870a41/lib/std/compress/flate/CircularBuffer.zig#L65) 8 times for cur_len of 1,2,4,8,16,32,64,128 and one copy out of loop for remaining 3 bytes. And that repeats 4 million times. That is all important work in this case.

The whole purpose of the loop is to ensure non overlapping buffers so we can use `@memcpy`.  But then replacing `@memcpy` with `copyForwards` gains performance. Loop can be replaces with single `copyForwards` but that ruins performance to 1500ms. 

Any ideas why, in this particular case, `copyForwards` outperforms `@memcpy`.