@ianic 2024-02-28
This numbers are from arm Linux (Linux vm on mac M1). But I got similar results on x86 Linux.