@andrewrk 2024-02-28
for example the memcpy implementation could be from glibc, from musl, or from zig's compiler-rt 

zig's compiler-rt memcpy has an open PR for optimizing it: https://github.com/ziglang/zig/pull/18912