@IntegratedQuantum 2024-02-29
> For this data (1G empty) if I replace `@memcpy` with `std.mem.copyForward` (as used in v1) I got better result 144ms instead of 215ms.

Interesting. I tried this but it made no performance difference for me. Did you maybe change something else as well?
Curiously though, using `std.mem.copyBackwards` does improve performance from 179 ms to 145 ms on my computer (x86_64 linux).

> zig's compiler-rt memcpy has an open PR for optimizing it: https://github.com/ziglang/zig/pull/18912

I also tried to apply this patch to memcpy, and it does give a small improvement, from 179 ms to 165 ms. Still far from v1's performance though.