@ianic 2024-02-29
> Did you maybe change something else as well?

No, but I must correct myself after looking once more, I'm finding this difference only on arm not  on x86.

Here are benchmarks from arm Linux:

Benchmark 1 - copyForwards
Benchmark 2 - @memcpy
Benchmark 3 - using same approach as v1 

This is from arm:
```
Benchmark 1: zig-out-copy-forwards/bin/flate_bench_std -i 4
  Time (mean ± σ):     150.8 ms ±   0.6 ms    [User: 150.5 ms, System: 0.3 ms]
  Range (min … max):   149.7 ms … 151.7 ms    19 runs

Benchmark 2: zig-out-std/bin/flate_bench_std -i 4
  Time (mean ± σ):     232.7 ms ±   0.9 ms    [User: 232.1 ms, System: 0.3 ms]
  Range (min … max):   231.9 ms … 235.2 ms    12 runs

  Warning: The first benchmarking run for this command was significantly slower than the rest (235.2 ms). This could be caused by (filesystem) caches that were not filled until after the first run. You should consider using the '--warmup' option to fill those caches before the actual benchmark. Alternatively, use the '--prepare' option to clear the caches before each timing run.

Benchmark 3: zig-out/bin/flate_bench_std -i 4
  Time (mean ± σ):     128.1 ms ±   1.1 ms    [User: 127.8 ms, System: 0.2 ms]
  Range (min … max):   127.1 ms … 131.6 ms    23 runs

Summary
  zig-out/bin/flate_bench_std -i 4 ran
    1.18 ± 0.01 times faster than zig-out-copy-forwards/bin/flate_bench_std -i 4
    1.82 ± 0.02 times faster than zig-out-std/bin/flate_bench_std -i 4
```

Benchmark 3, which has v1 performance is using this code:
```Zig
        if (from_end < buffer_len and to_end < buffer_len) {
            while (to < to_end) {
                const dst = self.buffer[to..to_end];
                const src = self.buffer[from..to];

                to += brk: {
                    if (src.len > dst.len) {
                        std.mem.copyForwards(u8, dst, src[0..dst.len]);
                        break :brk dst.len;
                    }
                    std.mem.copyForwards(u8, dst[0..src.len], src);
                    break :brk src.len;
                };
            }
       return;
      }
```
but it also ruins performance in other tests so I'm not considering including it in our implementation. 

This case (empty file) is interesting but I don't think it is highly relevant. You also mentioned "some uncompressed images with a small palette" do you have any recommendation from where to download some relevant dataset?