@IntegratedQuantum 2024-02-29
> Benchmark 3, which has v1 performance is using this code

Just tried this and it's still slower than v1 for me on x86, but certainly faster. Not sure what's going on there.

> This case (empty file) is interesting but I don't think it is highly relevant. You also mentioned "some uncompressed images with a small palette" do you have any recommendation from where to download some relevant dataset?

I can just send you some data: [data_sets.tar.gz](https://github.com/ziglang/zig/files/14451035/data_sets.tar.gz)

First one is the file that I meant by "some uncompressed images with a small palette". It's ~100 pixel art animation frames(192×108) stored using an 8-bit palette inside a custom format from a self-written pixel art editor.
Varies between 428 µs with this PR and 410 µs with your benchmark 3 (insignificant difference)

Second one is a screenshot I just now took of VSCode converted to bitmap with GIMP. Should be fairly good to trigger this issue because the background is mostly uniformly gray in sections longer than 258 bytes.
Varies between 5.7 ms with this PR and 5.2 ms with your benchmark 3 (10% difference)