@andrewrk 2024-02-29
Reminder that binary layout changes nondeterministically every time you compile, and can have drastic impacts on performance. [Related talk](https://www.youtube.com/watch?v=r-TLSBdHe1A).