@ianic 2024-02-29
I made a [project](https://github.com/ianic/zig-flate-benchmark) to ease benchmarking and experimenting with various improvements in flate.   
I put there few 'relevant' test cases, including your images and 1G empty file.